added ft_strnstr to ft_strdup, completed the mandatory part

+++ core/libft/ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: joho <joho@student.42singapore.sg>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/02 13:07:22 by joho              #+#    #+#             */
-/*   Updated: 2025/06/02 15:29:46 by joho             ###   ########.fr       */
+/*   Updated: 2025/06/03 16:58:58 by joho             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
